+++ packages/horizon-collector/src/horizon_collector/mapping/__init__.py
@@ -1,0 +1,5 @@
+"""Observation mapping."""
+
+from horizon_collector.mapping.mapper import CatalogObservationMapper, ObservationSourceMapping
+
+__all__ = ["CatalogObservationMapper", "ObservationSourceMapping"]